Create ros_arduino_publisher_subscriber.py
In this file, I have prepared a python code for arduino in which I shall prepare a node which will acquire some data from sensor (such as position sensor) using Arduino and then publish it. Then, I shall write another node, which will subscribe to that node and then based on the sensor data, it will publish something for example pwm for the motor.